requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
python-dotenv==1.0.1
google-genai==0.3.0
schedule==1.2.1
tavily-python==0.5.0
duckduckgo-search>=6.0.0
